# src/main/resources/api.yml
openapi: 3.0.0
info:
  title: Campus Wall API
  version: "1.0.0"
  description: API for campus social features
servers:
  - url: /api/v1
...
paths:
...
  /posts/{postId}:
    get:
      tags: [posts]
      summary: Get post by ID
      description: Retrieve a single post by its ID
      operationId: getPostById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
          description: ID of the post to retrieve
      responses:
        "200":
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDTO"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - no access to this post
        "404":
          description: Post not found
  /posts/{postId}/comments:
    post:
      tags: [posts]
      summary: Add comment to post
      description: Add a comment to an existing post
      operationId: addComment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
          description: ID of the post to comment on
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentDTO"
        "400":
          description: Invalid request data
        "401":
          description: Unauthorized - authentication required
        "404":
          description: Post not found
...
  /posts/{postId}/hide:
    patch:
      tags: [posts]
      summary: Hide a post
      description: Hide a post (soft-delete). Only the post author can hide their own post. When a post is hidden, all its comments are also hidden.
      operationId: hidePost
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
          description: ID of the post to hide
      responses:
        "200":
          description: Post hidden successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post hidden successfully"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - only post author can hide their own post
        "404":
          description: Post not found
  /posts/{postId}/unhide:
    patch:
      tags: [posts]
      summary: Unhide a post
      description: Unhide a previously hidden post (undo soft-delete). Only the post author can unhide their own post. When a post is unhidden, all its comments are restored.
      operationId: unhidePost
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
          description: ID of the post to unhide
      responses:
        "200":
          description: Post unhidden successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post unhidden successfully"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - only post author can unhide their own post
        "404":
          description: Post not found
...
  /admin/posts:
    get:
      tags: [admin]
      summary: Get all posts
      description: Admin endpoint to retrieve all posts with pagination, filters, and sorting
      operationId: adminGetPosts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: userId
          required: false
          schema:
            type: string
            format: uuid
          description: Filter posts by user ID
        - in: query
          name: hidden
          required: false
          schema:
            type: boolean
          description: Filter by hidden status
        - in: query
          name: sortBy
          required: false
          schema:
            type: string
            enum: [createdAt, likeCount, commentCount, userId]
          description: Field to sort by
        - in: query
          name: sortOrder
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order (ascending or descending)
      responses:
        "200":
          description: Posts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminPostDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
  /admin/posts/{id}:
    get:
      tags: [admin]
      summary: Get a post by ID
      description: Admin endpoint to retrieve a specific post by its ID
      operationId: adminGetPostById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Post ID to retrieve
      responses:
        "200":
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPostDTO"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
        "404":
          description: Post not found
    delete:
      tags: [admin]
      summary: Delete a post
      description: Admin endpoint to soft-delete (hide) a post
      operationId: adminDeletePost
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Post ID to delete
      responses:
        "200":
          description: Post deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post deleted successfully"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
        "404":
          description: Post not found
  /admin/posts/{id}/hide:
    put:
      tags: [admin]
      summary: Hide a post
      operationId: adminHidePost
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Post hidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/posts/{id}/unhide:
    put:
      tags: [admin]
      summary: Unhide a post
      operationId: adminUnhidePost
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Post unhidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/posts/{id}/comments:
    get:
      tags: [admin]
      summary: Get comments for a post
      operationId: adminGetPostComments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Post comments retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminCommentDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/posts/by-wall:
    get:
      tags: [admin]
      summary: Get posts by wall type with sorting
      description: Admin endpoint to retrieve posts filtered by wall type (national/campus/all) with sorting. Does NOT filter by schoolDomain and includes both hidden and non-hidden posts.
      operationId: adminGetPostsByWall
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: wall
          required: false
          schema:
            type: string
            enum: [national, campus]
          description: Filter posts by wall type. If null, returns all posts.
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: sortBy
          required: false
          schema:
            type: string
            enum: [NEWEST, OLDEST, MOST_LIKED, LEAST_LIKED]
            default: NEWEST
          description: Sort order for posts
      responses:
        "200":
          description: Posts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminPostDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
...
components:
  schemas:
...
    PostDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          description: Post title
        content:
          type: string
        wall:
          type: string
          enum: [campus, national]
          description: Which wall the post is on
        author:
          type: object
          properties:
            id:
              type: string
            profileName:
              type: string
              description: Profile name of the post author
            isAnonymous:
              type: boolean
              description: Whether author is anonymous
        likes:
          type: integer
          description: Number of likes
        comments:
          type: integer
          description: Number of comments
        liked:
          type: boolean
          description: Whether current user has liked this post
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCommentRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          description: Comment text content
    CommentDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
          description: ID of the parent entity (post, internship, or marketplace item)
        parentType:
          type: string
          enum: [POST, INTERNSHIP, MARKETPLACE]
          description: Type of the parent entity
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        author:
          type: object
          properties:
            id:
              type: string
            profileName:
              type: string
              description: Profile name of the comment author
            isAnonymous:
              type: boolean

    # ----- Internships -----
...
    AdminPostDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        profileName:
          type: string
        title:
          type: string
        content:
          type: string
        wall:
          type: string
          enum: [campus, national]
        schoolDomain:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        hidden:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminCommentDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        profileName:
          type: string
        text:
          type: string
        hidden:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT